reuse the code move the initialize function around
